refactor: migrate to Astro Content Layer API, update component types, and remove incompatible wrangler config
