+++ scripts/fix-routes.mjs
@@ -9,7 +9,7 @@
 //   - /blog/opal      → include /* AND exclude /blog/opal/* → exclude wins → CDN ✓
 //   - /es/blog/opal   → include /* AND exclude /es/blog/opal/* → exclude wins → CDN ✓
 
-import { writeFileSync, readFileSync, existsSync, readdirSync, statSync } from "fs"
+import { writeFileSync, existsSync, rmSync, readdirSync, statSync } from "fs"
 import { join } from "path"
 
 function getStaticDirs(dir) {
@@ -69,29 +69,12 @@ if (totalRules > 100) {
 	process.exit(1)
 }
 
-// Fix dist/server/wrangler.json for Cloudflare Pages compatibility.
-// @astrojs/cloudflare generates a wrangler.json with fields Pages rejects.
+// Remove dist/server/wrangler.json — it is a Worker config generated by
+// @astrojs/cloudflare that is incompatible with Cloudflare Pages. Pages
+// auto-detects and tries to use it, conflicting with the root wrangler.toml.
+// The root wrangler.toml (pages_build_output_dir) is the only config needed.
 const wranglerPath = "dist/server/wrangler.json"
 if (existsSync(wranglerPath)) {
-	const wrangler = JSON.parse(readFileSync(wranglerPath, "utf-8"))
-
-	// Pages rejects triggers: {} — must be omitted when there are no crons
-	if (wrangler.triggers && Object.keys(wrangler.triggers).length === 0) {
-		delete wrangler.triggers
-	}
-
-	// KV namespace bindings without an "id" can't be deployed; remove them
-	if (Array.isArray(wrangler.kv_namespaces)) {
-		wrangler.kv_namespaces = wrangler.kv_namespaces.filter((ns) => ns.id)
-		if (wrangler.kv_namespaces.length === 0) delete wrangler.kv_namespaces
-	}
-
-	// "ASSETS" is reserved in Pages projects; Pages provides it automatically
-	if (wrangler.assets?.binding === "ASSETS") {
-		delete wrangler.assets.binding
-		if (Object.keys(wrangler.assets).length === 0) delete wrangler.assets
-	}
-
-	writeFileSync(wranglerPath, JSON.stringify(wrangler, null, 2))
-	console.log("[postbuild] dist/server/wrangler.json fixed ✓")
+	rmSync(wranglerPath)
+	console.log("[postbuild] dist/server/wrangler.json removed ✓")
 }
