--- src/content/config.ts
@@ -1,46 +1,0 @@
-import { defineCollection } from "astro:content"
-import { z } from "astro/zod"
-import { CATEGORY_SLUGS } from "../config/categories"
-
-const projectsCollection = defineCollection({
-	type: "content",
-})
-const proyectosCollection = defineCollection({
-	type: "content",
-})
-const blogCollection = defineCollection({
-	type: "content",
-	schema: z.object({
-		title: z.string(),
-		description: z.string(),
-		search_context: z.string(),
-		date: z.string(),
-		lang: z.enum(["en", "es"]),
-		image: z.string(),
-		og_image: z.string().optional(),
-		categories: z.array(z.enum(CATEGORY_SLUGS)),
-		status: z.enum(["active", "draft"]),
-	}),
-})
-
-const seoCollection = defineCollection({
-	type: "data",
-	schema: z.object({
-		pages: z.record(
-			z.string(),
-			z.object({
-				title: z.string(),
-				description: z.string(),
-				og_title: z.string().optional(),
-				og_image: z.string().optional(),
-			})
-		),
-	}),
-})
-
-export const collections = {
-	projects: projectsCollection,
-	proyectos: proyectosCollection,
-	blog: blogCollection,
-	seo: seoCollection,
-}
